LayoutTests/fast/css-grid-layout/grid-repeat-calc.html

<!DOCTYPE html>
<body>
<script src="../../resources/js-test-pre.js"></script>
<div id="testDiv" style="width: 400px; height: 400px"></div>
<script>
description("Tests using calculated value in function() for 'grid-template-rows' CSS property works.");

var testDiv = document.getElementById("testDiv");
//grid-template-rows: repeat(1, 18px);
// grid-template-columns: repeat(1, 15%);
shouldBeEmptyString("testDiv.style['grid-template-rows']");
evalAndLog("testDiv.style['grid-template-rows'] = 'repeat(calc(1 + 1), 18px)'");
shouldBeEqualToString("testDiv.style['grid-template-rows']", "18px 18px");
shouldBeEqualToString("window.getComputedStyle(testDiv).getPropertyValue('grid-template-rows')", "18px 18px");

</script>
<script src="../../resources/js-test-post.js"></script>
</body>
